+++ src/fields/Categories.php
@@ -8,6 +8,7 @@ use craft\base\Element as BaseElement;
 use craft\elements\Category as CategoryElement;
 use craft\feedme\base\Field;
 use craft\feedme\base\FieldInterface;
+use craft\feedme\helpers\DataHelper;
 use craft\feedme\Plugin;
 use craft\fields\Categories as CategoriesField;
 use craft\helpers\Db;
@@ -59,6 +60,17 @@ class Categories extends Field implements FieldInterface
         $value = $this->fetchArrayValue();
         $default = $this->fetchDefaultArrayValue();
 
+        // if the mapped value is not set in the feed
+        if ($value === null) {
+            return null;
+        }
+
+        // if value from the feed is empty and default is not set
+        // return an empty array; no point bothering further
+        if (empty($default) && DataHelper::isArrayValueEmpty($value)) {
+            return [];
+        }
+
         $source = Hash::get($this->field, 'settings.source');
         $branchLimit = Hash::get($this->field, 'settings.branchLimit');
         $targetSiteId = Hash::get($this->field, 'settings.targetSiteId');
@@ -67,16 +79,13 @@ class Categories extends Field implements FieldInterface
         $create = Hash::get($this->fieldInfo, 'options.create');
         $fields = Hash::get($this->fieldInfo, 'fields');
         $node = Hash::get($this->fieldInfo, 'node');
+        $nodeKey = null;
 
         // Get source id's for connecting
         [, $groupUid] = explode(':', $source);
         $groupId = Db::idByUid('{{%categorygroups}}', $groupUid);
 
         $foundElements = [];
-
-        if (!$value) {
-            return $foundElements;
-        }
 
         foreach ($value as $dataValue) {
             // Prevent empty or blank values (string or array), which match all elements
@@ -92,7 +101,7 @@ class Categories extends Field implements FieldInterface
 
             // special provision for falling back on default BaseRelationField value
             // https://github.com/craftcms/feed-me/issues/1195
-            if (empty($dataValue)) {
+            if (trim($dataValue) === '') {
                 $foundElements = $default;
                 break;
             }
@@ -139,6 +148,8 @@ class Categories extends Field implements FieldInterface
             if ((count($ids) == 0) && $create && $match === 'title') {
                 $foundElements[] = $this->_createElement($dataValue, $groupId);
             }
+
+            $nodeKey = $this->getArrayKeyFromNode($node);
         }
 
         // Check for field limit - only return the specified amount
@@ -148,7 +159,7 @@ class Categories extends Field implements FieldInterface
 
         // Check for any sub-fields for the element
         if ($fields) {
-            $this->populateElementFields($foundElements);
+            $this->populateElementFields($foundElements, $nodeKey);
         }
 
         $foundElements = array_unique($foundElements);