+++ src/fields/Matrix.php
@@ -5,6 +5,7 @@ namespace craft\feedme\fields;
 use Cake\Utility\Hash;
 use craft\feedme\base\Field;
 use craft\feedme\base\FieldInterface;
+use craft\feedme\helpers\DataHelper;
 use craft\feedme\Plugin;
 use craft\fields\Matrix as MatrixField;
 
@@ -62,6 +63,7 @@ class Matrix extends Field implements FieldInterface
         foreach ($this->feedData as $nodePath => $value) {
             // Get the field mapping info for this node in the feed
             $fieldInfo = $this->_getFieldMappingInfoForNodePath($nodePath, $blocks);
+            $nodePathSegments = explode('/', $nodePath);
 
             // If this is data concerning our Matrix field and blocks
             if ($fieldInfo) {
@@ -70,20 +72,7 @@ class Matrix extends Field implements FieldInterface
                 $subFieldInfo = $fieldInfo['subFieldInfo'];
                 $isComplexField = $fieldInfo['isComplexField'];
 
-                $nodePathSegments = explode('/', $nodePath);
-                $blockIndex = Hash::get($nodePathSegments, 1);
-
-                if (!is_numeric($blockIndex)) {
-                    // Try to check if its only one-level deep (only importing one block type)
-                    // which is particularly common for JSON.
-                    $blockIndex = Hash::get($nodePathSegments, 2);
-
-                    if (!is_numeric($blockIndex)) {
-                        $blockIndex = 0;
-                    }
-                }
-
-                $key = $blockIndex . '.' . $blockHandle . '.' . $subFieldHandle;
+                $key = $this->_getBlockKey($nodePathSegments, $blockHandle, $subFieldHandle);
 
                 // Check for complex fields (think Table, Super Table, etc), essentially anything that has
                 // sub-fields, and doesn't have data directly mapped to the field itself. It needs to be
@@ -109,6 +98,18 @@ class Matrix extends Field implements FieldInterface
                     $fieldData[$key] = $parsedValue;
                 }
             }
+
+            foreach ($blocks as $blockHandle => $fields) {
+                foreach ($fields['fields'] as $fieldHandle => $fieldInfo) {
+                    $node = Hash::get($fieldInfo, 'node');
+                    if ($node === 'usedefault') {
+                        $key = $this->_getBlockKey($nodePathSegments, $blockHandle, $fieldHandle);
+
+                        $parsedValue = DataHelper::fetchSimpleValue($this->feedData, $fieldInfo);
+                        $fieldData[$key] = $parsedValue;
+                    }
+                }
+            }
         }
 
         // Handle some complex fields that don't directly have nodes, but instead have nested properties mapped.
@@ -124,10 +125,12 @@ class Matrix extends Field implements FieldInterface
 
             $parsedValue = $this->_parseSubField($nodePaths, $subFieldHandle, $subFieldInfo);
 
-            if (isset($fieldData[$key])) {
-                $fieldData[$key] = array_merge_recursive($fieldData[$key], $parsedValue);
-            } else {
-                $fieldData[$key] = $parsedValue;
+            if ($parsedValue !== null) {
+                if (isset($fieldData[$key])) {
+                    $fieldData[$key] = array_merge_recursive($fieldData[$key], $parsedValue);
+                } else {
+                    $fieldData[$key] = $parsedValue;
+                }
             }
         }
 
@@ -145,13 +148,16 @@ class Matrix extends Field implements FieldInterface
             $subFieldHandle = $handles[2];
 
             $disabled = Hash::get($this->fieldInfo, 'blocks.' . $blockHandle . '.disabled', false);
-            $collapsed = Hash::get($this->fieldInfo, 'blocks.' . $blockHandle . '.collapsed', false);
 
             // Prepare an array that's ready for Matrix to import it
             $preppedData[$blockIndex . '.type'] = $blockHandle;
             $preppedData[$blockIndex . '.enabled'] = !$disabled;
-            $preppedData[$blockIndex . '.collapsed'] = $collapsed;
             $preppedData[$blockIndex . '.fields.' . $subFieldHandle] = $value;
+        }
+
+        // if there's nothing in the prepped data, return null, as if mapping doesn't exist
+        if (empty($preppedData)) {
+            return null;
         }
 
         $expanded = Hash::expand($preppedData);
@@ -160,7 +166,15 @@ class Matrix extends Field implements FieldInterface
         $index = 1;
         $resultBlocks = [];
         foreach ($expanded as $blockData) {
-            $resultBlocks['new' . $index++] = $blockData;
+            // all the fields are empty and setEmptyValues is off, ignore the block
+            if (
+                !empty(array_filter(
+                    $blockData['fields'],
+                    fn($value) => (is_string($value) && !empty($value)) || (is_array($value) && !empty(array_filter($value)))
+                ))
+            ) {
+                $resultBlocks['new' . $index++] = $blockData;
+            }
         }
 
         return $resultBlocks;
@@ -171,9 +185,34 @@ class Matrix extends Field implements FieldInterface
     // =========================================================================
 
     /**
+     * Get block's key
+     *
+     * @param array $nodePathSegments
+     * @param string $blockHandle
+     * @param string $fieldHandle
+     * @return string
+     */
+    private function _getBlockKey(array $nodePathSegments, string $blockHandle, string $fieldHandle): string
+    {
+        $blockIndex = Hash::get($nodePathSegments, 1);
+
+        if (!is_numeric($blockIndex)) {
+            // Try to check if its only one-level deep (only importing one block type)
+            // which is particularly common for JSON.
+            $blockIndex = Hash::get($nodePathSegments, 2);
+
+            if (!is_numeric($blockIndex)) {
+                $blockIndex = 0;
+            }
+        }
+
+        return $blockIndex . '.' . $blockHandle . '.' . $fieldHandle;
+    }
+
+    /**
      * @param $nodePath
      * @param $blocks
-     * @return array|null
+     * @return array|null|string
      */
     private function _getFieldMappingInfoForNodePath($nodePath, $blocks): ?array
     {
@@ -202,7 +241,7 @@ class Matrix extends Field implements FieldInterface
                     }
                 }
 
-                if ($feedPath == $node || $node === 'usedefault') {
+                if ($feedPath == $node) {
                     return [
                         'blockHandle' => $blockHandle,
                         'subFieldHandle' => $subFieldHandle,