+++ src/fields/TypedLink.php
@@ -57,6 +57,14 @@ class TypedLink extends Field implements FieldInterface
             $preppedData[$subFieldHandle] = DataHelper::fetchValue($this->feedData, $subFieldInfo, $this->feed);
         }
 
+        if (empty(
+        array_filter($preppedData, function($val) {
+            return $val !== null;
+        })
+        )) {
+            return null;
+        }
+
         // Protect against sending an empty array
         if (!$preppedData) {
             return null;