+++ src/web/twig/variables/FeedMeVariable.php
@@ -17,6 +17,7 @@ use craft\fields\PlainText;
 use craft\fields\RadioButtons;
 use craft\fields\Url;
 use craft\helpers\DateTimeHelper;
+use craft\helpers\Html;
 use craft\helpers\UrlHelper;
 use craft\models\CategoryGroup;
 use craft\models\Section;
@@ -93,7 +94,7 @@ class FeedMeVariable extends ServiceLocator
                     continue;
                 }
 
-                $values[$value[$index]] = $value[$label];
+                $values[$value[$index]] = Html::encode($value[$label]);
             }
         }
 
@@ -344,6 +345,8 @@ class FeedMeVariable extends ServiceLocator
             Number::class,
             PlainText::class,
             RadioButtons::class,
+            'craft\ckeditor\Field',
+            'craft\redactor\Field',
         ];
 
         return in_array($class, $supportedSubFields, true);